requirements.txt
azure-identity
azure-purview-scanning
azure-purview-administration
azure-purview-catalog
azure-purview-account
azure-core

